+++ lib/elasticity/job_flow.rb
@@ -18,7 +18,7 @@ module Elasticity
     def initialize(access, secret)
       @action_on_failure = 'TERMINATE_JOB_FLOW'
       @ec2_key_name = 'default'
-      @hadoop_version = '0.20'
+      @hadoop_version = '0.20.205'
       @instance_count = 2
       @master_instance_type = 'm1.small'
       @name = 'Elasticity Job Flow'